+++ src/CMakeLists.txt
@@ -1,0 +1,5 @@
+cmake_minimum_required(VERSION 3.5)
+
+add_subdirectory(ui)
+add_subdirectory(models)
+
