+++ src/models/CMakeLists.txt
@@ -1,0 +1,4 @@
+cmake_minimum_required(VERSION 3.5)
+
+add_subdirectory(filemodel)
+
