+++ src/models/filemodel/filemodel.cpp
@@ -1,0 +1,6 @@
+#include "filemodel.h"
+
+FileModel::FileModel()
+{
+
+}
