+++ src/ui/CMakeLists.txt
@@ -1,0 +1,3 @@
+cmake_minimum_required(VERSION 3.5)
+
+add_subdirectory(mainform)
