+++ src/ui/mainform/mainform.h
@@ -3,6 +3,8 @@
 
 #include <QWidget>
 #include <QQmlApplicationEngine>
+#include <QtQml>
+#include <QFileSystemModel>
 
 class MainForm : public QObject
 {
@@ -17,5 +19,6 @@ public:
 private:
     QQmlApplicationEngine* qml_app_engine;
     QObject* qml_app_root_object;
+    QFileSystemModel *model;
 };
 #endif // MAINFORM_H
